Large amount of work on new features in Fallout 4
Type names are now mapped to pretty names, with appropriate casing.
The casing of the most keywords is now consistently PascalCase.
Output PEX format info in decompilation header.
Structs are now decompiled, and their members placed in the order in the debug info, if the info is available.
Properties are now properly sorted into the groups and order present in the debug info, if available.
Const variables will be output as const.
Only output functions/states/structs/properties splitter comments if there are actually any of them present.
The PropertyGroup debug info is now appropriately marked as a DocumentedItem.
Float values are now only written with the minimum number of decimal places required, or no decimals at all if the value is integral.
Update version number to 1.0.4.